+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod engine;
 
 pub mod diode;
 pub mod isine;
+pub mod vsine;
 
 pub mod capacitor;