+++ tests/test_dc_bridge_p_loaded.rs
@@ -45,8 +45,8 @@ fn build_v() -> Circuit {
     let mut ckt = Circuit::new();
 
     // bridge input voltage
-    ckt.elements.push(Element::V(VoltageSource{p: 1, n: 2, value: 10.0}));
-    ckt.elements.push(Element::V(VoltageSource{p: 2, n: 0, value: 0.0}));
+    ckt.elements.push(Element::V(VoltageSource{p: 1, n: 2, value: 10.0, idx:0 }));
+    ckt.elements.push(Element::V(VoltageSource{p: 2, n: 0, value: 0.0 , idx:1 }));
 
     // Diode bridge
     //  (1) is top
@@ -67,7 +67,7 @@ fn build_vv() -> Circuit {
     let mut ckt = Circuit::new();
 
     // bridge input voltage
-    ckt.elements.push(Element::V(VoltageSource{p: 1, n: 0, value: 10.0}));
+    ckt.elements.push(Element::V(VoltageSource{p: 1, n: 0, value: 10.0, idx:0}));
 
     // Diode bridge
     //  (1) is top