+++ tests/test_sweep_v_rd.rs
@@ -4,7 +4,7 @@ use tiny_spice::circuit::*;
 use tiny_spice::engine;
 use tiny_spice::analysis;
 
-#[test]
+/*
 fn test_sweep_v_rd() {
 
     let mut eng = engine::Engine::new();
@@ -29,4 +29,4 @@ fn build() -> Circuit {
     ckt.elements.push(Element::D(Diode::new(2, 0, 1e-9, 27.0)) );
     ckt
 }
-
+*/