+++ tests/test_v_d_vs_d.rs
@@ -27,7 +27,7 @@ fn test_v_d_vs_d() {
 // (1) V --> D1 (fwd) (2) --> D2(rev) (0)
 fn build() -> Circuit {
     let mut ckt = Circuit::new();
-    ckt.elements.push( Element::V(VoltageSource{p: 1, n: 0, value: 5.0}) );
+    ckt.elements.push( Element::V(VoltageSource{p: 1, n: 0, value: 5.0, idx:0}) );
     // 'forward' biased
     ckt.elements.push( Element::D(Diode::new(1, 2, 1e-9, 27.0)) );
     // reversed bias