+++ tests/test_vvrrr.rs
@@ -8,7 +8,7 @@ mod common;
 use crate::common::assert_nearly;
 
 #[test]
-fn test() {
+fn test_vvrrr() {
 
     let mut eng = engine::Engine::new();
     let mut cfg = analysis::Configuration::new();
@@ -43,7 +43,7 @@ fn build() -> circuit::Circuit {
         circuit::Element::R(circuit::Resistor{a: 0, b: 1, value: 2.0}),
     );
     ckt.elements.push(
-        circuit::Element::V(circuit::VoltageSource{p: 2, n: 1, value: 32.0}),
+        circuit::Element::V(circuit::VoltageSource{p: 2, n: 1, value: 32.0, idx: 0}),
     );
     ckt.elements.push(
         circuit::Element::R(circuit::Resistor{a: 2, b: 3, value: 4.0}),
@@ -52,7 +52,7 @@ fn build() -> circuit::Circuit {
         circuit::Element::R(circuit::Resistor{a: 2, b: 0, value: 8.0}),
     );
     ckt.elements.push(
-        circuit::Element::V(circuit::VoltageSource{p: 3, n: 0, value: 20.0}),
+        circuit::Element::V(circuit::VoltageSource{p: 3, n: 0, value: 20.0, idx: 1}),
     );
     ckt
 }